pyproject.toml
[tool.poetry]
name = "scorecard-tracing-demo"
version = "0.1.0"
description = ""
authors = ["Scorecard AI <team@getscorecard.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
openai = "^1.21.2"
openinference-instrumentation-openai = "^0.1.4"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
opentelemetry-exporter-otlp-proto-http = "^1.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
